pyproject.toml
[project]
name = "bot-windows"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "pytelegrambotapi>=4.27.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "ujson>=5.10.0",
]
